requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
Pillow
django-filter